Split x86 register handling into separate module
