+++ include/regalloc.h
@@ -15,10 +15,5 @@ void regalloc_run(ir_builder_t *ir, regalloc_t *ra);
 /* Free resources held by the allocator */
 void regalloc_free(regalloc_t *ra);
 
-/* Return physical register name for given index */
-const char *regalloc_reg_name(int idx);
-
-/* Select register naming for 64-bit or 32-bit mode. */
-void regalloc_set_x86_64(int enable);
 
 #endif /* VC_REGALLOC_H */
